Refactor: extract CiiParser as base for FacturXParser
The CII (Cross Industry Invoice) parsing logic now lives in CiiParser,
usable directly for any CrossIndustryInvoice XML (FacturX, ZUGFeRD 2.1.1,
XRechnung). FacturXParser becomes a named empty subclass for callers that
want to type-hint on the specific standard. InvoiceReader auto-detection
now instantiates CiiParser for CrossIndustryInvoice documents.
